+++ src/work_mgga.c
@@ -24,7 +24,7 @@
  * @param[in,out] func_type: pointer to functional structure
  */
 static void 
-work_mgga(const XC(func_type) *p, int np,
+work_mgga(const XC(func_type) *p, size_t np,
          const double *rho, const double *sigma, const double *lapl, const double *tau,
           double *zk, MGGA_OUT_PARAMS_NO_EXC(double *))
 {
